backend/requirements.txt
fastapi[standard]
passlib[bcrypt]
python-jose
pydantic
